+++ examples/demo-rollup/src/main.rs
@@ -11,6 +11,7 @@ use sov_demo_rollup::{
     MockDemoRollup, MockNomtDemoRollup,
 };
 use sov_mock_da::storable::service::StorableMockDaService;
+use sov_modules_api::capabilities::RollupHeight;
 use sov_modules_api::execution_mode::Native;
 use sov_modules_rollup_blueprint::logging::initialize_logging;
 use sov_modules_rollup_blueprint::{FullNodeBlueprint, Rollup};
@@ -44,6 +45,10 @@ struct Args {
     /// Listen address for Prometheus exporter.
     #[arg(long, default_value = "127.0.0.1:9845")]
     prometheus_exporter_bind: String,
+
+    /// Stops the rollup at a given height.
+    #[arg(long, default_value = None)]
+    stop_at_rollup_height: Option<u64>,
 }
 
 #[derive(clap::ValueEnum, Clone, Debug)]
@@ -87,6 +92,7 @@ async fn run() -> anyhow::Result<()> {
         "Running demo rollup with prover config"
     );
 
+    let stop_at_rollup_height = args.stop_at_rollup_height.map(RollupHeight::new);
     match (args.da_layer, args.storage) {
         (SupportedDaLayer::Mock, SupportedStorage::Jmt) => {
             let prover_config = prover_config_disc
@@ -95,6 +101,7 @@ async fn run() -> anyhow::Result<()> {
                 &GenesisPaths::from_dir(&args.genesis_config_dir),
                 rollup_config_path,
                 prover_config,
+                args.stop_at_rollup_height.map(RollupHeight::new),
             )
             .await
             .context("Failed to initialize MockDa rollup")?;
@@ -107,6 +114,7 @@ async fn run() -> anyhow::Result<()> {
                 &GenesisPaths::from_dir(&args.genesis_config_dir),
                 rollup_config_path,
                 prover_config,
+                stop_at_rollup_height,
             )
             .await
             .context("Failed to initialize NOMT based MockDa rollup")?;
@@ -119,6 +127,7 @@ async fn run() -> anyhow::Result<()> {
                 &GenesisPaths::from_dir(&args.genesis_config_dir),
                 rollup_config_path,
                 prover_config,
+                stop_at_rollup_height,
             )
             .await
             .context("Failed to initialize Celestia rollup")?;
@@ -131,6 +140,7 @@ async fn run() -> anyhow::Result<()> {
                 &GenesisPaths::from_dir(&args.genesis_config_dir),
                 rollup_config_path,
                 prover_config,
+                stop_at_rollup_height,
             )
             .await
             .context("Failed to initialize Celestia rollup")?;
@@ -160,6 +170,7 @@ async fn new_rollup_with_celestia_da(
     rt_genesis_paths: &GenesisPaths,
     rollup_config_path: &str,
     prover_config: Option<RollupProverConfig<Risc0>>,
+    stop_at_rollup_height: Option<RollupHeight>,
 ) -> anyhow::Result<Rollup<CelestiaDemoRollup<Native>, Native>> {
     debug!(config_path = rollup_config_path, "Starting Celestia rollup");
 
@@ -173,7 +184,12 @@ async fn new_rollup_with_celestia_da(
 
     let celestia_rollup = CelestiaDemoRollup::<Native>::default();
     celestia_rollup
-        .create_new_rollup(rt_genesis_paths, rollup_config, prover_config)
+        .create_new_rollup(
+            rt_genesis_paths,
+            rollup_config,
+            prover_config,
+            stop_at_rollup_height,
+        )
         .await
 }
 
@@ -181,6 +197,7 @@ async fn new_rollup_with_celestia_da_and_nomt(
     rt_genesis_paths: &GenesisPaths,
     rollup_config_path: &str,
     prover_config: Option<RollupProverConfig<Risc0>>,
+    stop_at_rollup_height: Option<RollupHeight>,
 ) -> anyhow::Result<Rollup<CelestiaNomtDemoRollup<Native>, Native>> {
     debug!(config_path = rollup_config_path, "Starting Celestia rollup");
 
@@ -194,7 +211,12 @@ async fn new_rollup_with_celestia_da_and_nomt(
 
     let celestia_rollup = CelestiaNomtDemoRollup::<Native>::default();
     celestia_rollup
-        .create_new_rollup(rt_genesis_paths, rollup_config, prover_config)
+        .create_new_rollup(
+            rt_genesis_paths,
+            rollup_config,
+            prover_config,
+            stop_at_rollup_height,
+        )
         .await
 }
 
@@ -202,6 +224,7 @@ async fn new_rollup_with_mock_da_and_jmt(
     rt_genesis_paths: &GenesisPaths,
     rollup_config_path: &str,
     prover_config: Option<RollupProverConfig<Risc0>>,
+    stop_at_rollup_height: Option<RollupHeight>,
 ) -> anyhow::Result<Rollup<MockDemoRollup<Native>, Native>> {
     debug!(
         config_path = rollup_config_path,
@@ -218,7 +241,12 @@ async fn new_rollup_with_mock_da_and_jmt(
 
     let mock_rollup = MockDemoRollup::<Native>::default();
     mock_rollup
-        .create_new_rollup(rt_genesis_paths, rollup_config, prover_config)
+        .create_new_rollup(
+            rt_genesis_paths,
+            rollup_config,
+            prover_config,
+            stop_at_rollup_height,
+        )
         .await
 }
 
@@ -226,6 +254,7 @@ async fn new_rollup_with_mock_da_and_nomt(
     rt_genesis_paths: &GenesisPaths,
     rollup_config_path: &str,
     prover_config: Option<RollupProverConfig<Risc0>>,
+    stop_at_rollup_height: Option<RollupHeight>,
 ) -> anyhow::Result<Rollup<MockNomtDemoRollup<Native>, Native>> {
     debug!(
         config_path = rollup_config_path,
@@ -242,6 +271,11 @@ async fn new_rollup_with_mock_da_and_nomt(
 
     let mock_rollup = MockNomtDemoRollup::<Native>::default();
     mock_rollup
-        .create_new_rollup(rt_genesis_paths, rollup_config, prover_config)
+        .create_new_rollup(
+            rt_genesis_paths,
+            rollup_config,
+            prover_config,
+            stop_at_rollup_height,
+        )
         .await
 }